pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "taskmgr"
version = "0.1.0"
description = "Industry standard Django DRF boilerplate"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.0",
    "djangorestframework>=3.15",
    "django-environ>=0.11",
    "djangorestframework-simplejwt>=5.3",
    "drf-spectacular>=0.27",
    "django-cors-headers>=4.3",
    "psycopg2-binary>=2.9",
    "whitenoise>=6.6",
    "django-redis>=5.4",
    "redis>=5.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4",
    "pytest-django>=4.8",
    "pytest-cov>=5.0",
    "black>=24.4",
    "django-stubs>=4.2.7",
    "djangorestframework-stubs>=3.14.5",
    "mypy>=1.9.0",
]

[tool.mypy]
plugins = [
    "mypy_django_plugin.main",
    "mypy_drf_plugin.main",
]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "config.settings.local"

[tool.setuptools]
packages = ["apps", "config"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = ["E501"] # line too long (handled by formatter)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.local"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
